+++ projects/gtk-4.14.5/gtk/gtkbuilder.c
@@ -2301,15 +2301,15 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
 
   g_value_init (value, type);
 
-  switch (G_TYPE_FUNDAMENTAL (type))
+  switch ((uintptr_t)G_TYPE_FUNDAMENTAL (type))
     {
-    case G_TYPE_CHAR:
+    case (uintptr_t)G_TYPE_CHAR:
       g_value_set_schar (value, string[0]);
       break;
-    case G_TYPE_UCHAR:
+    case (uintptr_t)G_TYPE_UCHAR:
       g_value_set_uchar (value, (guchar)string[0]);
       break;
-    case G_TYPE_BOOLEAN:
+    case (uintptr_t)G_TYPE_BOOLEAN:
       {
         gboolean b;
 
@@ -2321,9 +2321,9 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
         g_value_set_boolean (value, b);
         break;
       }
-    case G_TYPE_INT:
-    case G_TYPE_LONG:
-    case G_TYPE_INT64:
+    case (uintptr_t)G_TYPE_INT:
+    case (uintptr_t)G_TYPE_LONG:
+    case (uintptr_t)G_TYPE_INT64:
       {
         gint64 l;
         char *endptr = NULL;
@@ -2348,9 +2348,9 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
           g_value_set_int64 (value, l);
         break;
       }
-    case G_TYPE_UINT:
-    case G_TYPE_ULONG:
-    case G_TYPE_UINT64:
+    case (uintptr_t)G_TYPE_UINT:
+    case (uintptr_t)G_TYPE_ULONG:
+    case (uintptr_t)G_TYPE_UINT64:
       {
         guint64 ul;
         char *endptr = NULL;
@@ -2374,7 +2374,7 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
           g_value_set_uint64 (value, ul);
         break;
       }
-    case G_TYPE_ENUM:
+    case (uintptr_t)G_TYPE_ENUM:
       {
         int enum_value;
         if (!_gtk_builder_enum_from_string (type, string, &enum_value, error))
@@ -2385,7 +2385,7 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
         g_value_set_enum (value, enum_value);
         break;
       }
-    case G_TYPE_FLAGS:
+    case (uintptr_t)G_TYPE_FLAGS:
       {
         guint flags_value;
 
@@ -2397,8 +2397,8 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
         g_value_set_flags (value, flags_value);
         break;
       }
-    case G_TYPE_FLOAT:
-    case G_TYPE_DOUBLE:
+    case (uintptr_t)G_TYPE_FLOAT:
+    case (uintptr_t)G_TYPE_DOUBLE:
       {
         double d;
         char *endptr = NULL;
@@ -2420,10 +2420,10 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
           g_value_set_double (value, d);
         break;
       }
-    case G_TYPE_STRING:
+    case (uintptr_t)G_TYPE_STRING:
       g_value_set_string (value, string);
       break;
-    case G_TYPE_VARIANT:
+    case (uintptr_t)G_TYPE_VARIANT:
       {
         GVariant *variant;
 
@@ -2434,7 +2434,7 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
           ret = FALSE;
       }
       break;
-    case G_TYPE_BOXED:
+    case (uintptr_t)G_TYPE_BOXED:
       if (G_VALUE_HOLDS (value, GDK_TYPE_RGBA))
         {
           GdkRGBA rgba = { 0 };
@@ -2538,8 +2538,8 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
           ret = FALSE;
         }
       break;
-    case G_TYPE_OBJECT:
-    case G_TYPE_INTERFACE:
+    case (uintptr_t)G_TYPE_OBJECT:
+    case (uintptr_t)G_TYPE_INTERFACE:
       if (G_VALUE_HOLDS (value, GDK_TYPE_PAINTABLE) ||
           G_VALUE_HOLDS (value, GDK_TYPE_TEXTURE))
         {
@@ -2727,7 +2727,7 @@ gtk_builder_value_from_string_type (GtkBuilder   *builder,
             }
         }
       break;
-    case G_TYPE_POINTER:
+    case (uintptr_t)G_TYPE_POINTER:
       if (G_VALUE_HOLDS (value, G_TYPE_GTYPE))
         {
           GType resolved_type;
